feat: Implementei algumas validações robustas e refatora funcionalidades
Novas Funcionalidades e Validações

- **Validação de Usuário:** Implementei um sistema robusto de validações para a criação e atualização de usuários:
    - **Email:** Restringe o cadastro a domínios de provedores populares (ex: @gmail.com, @hotmail.com).
    - **Senha:** Impõe uma política de complexidade (mín. 6 caracteres, com maiúscula, minúscula, número e símbolo).
    - **Telefone:** Garante que o número siga o padrão de São Paulo (DDD 11).
    - **CPF:** Adiciona a validação do cálculo dos dígitos verificadores.
    - **Atributos Físicos:** Define limites de mínimo e máximo para idade, peso e altura.
    - **Foto de Perfil:** Limita o tamanho do upload a 30 MB.

- **Renovação de Plano por Admin:** A funcionalidade de renovar planos Gold agora é uma responsabilidade exclusiva de administradores através do novo endpoint . O endpoint de autoatendimento foi removido.

**Correções de Bugs**

- **Tipo de Dado da Altura:** Corrige o campo  de  para  em toda a aplicação.

**Refatoração e Qualidade de Código**

- **Mensagem de Erro Unificada:** Centraliza a verificação de email, CPF e telefone duplicados no , provendo uma única mensagem de erro amigável.
- **Padronização de DTOs:** Remove anotações de persistência () dos DTOs, reforçando a separação de responsabilidades, e limpa comentários desnecessários para melhorar a legibilidade.

+++ src/main/java/br/com/saudeemacao/api/dto/UsuarioPerfilDTO.java
@@ -1,6 +1,6 @@
 package br.com.saudeemacao.api.dto;
 
-import br.com.saudeemacao.api.model.*;
+import br.com.saudeemacao.api.model.Usuario;
 import br.com.saudeemacao.api.model.EnumUsuario.ENivelAtividade;
 import br.com.saudeemacao.api.model.EnumUsuario.EPerfil;
 import br.com.saudeemacao.api.model.EnumUsuario.EPlano;
@@ -25,12 +25,9 @@ public class UsuarioPerfilDTO {
     private EPlano plano;
     private LocalDateTime dataUltimoTreino;
 
-    // =================================================================
-    // === NOVOS CAMPOS - Retornados se o aluno for GOLD
-    // =================================================================
     private Integer idade;
     private Integer peso;
-    private Integer altura;
+    private Double altura; // <-- ALTERAÇÃO AQUI: De Integer para Double
     private String objetivo;
     private ENivelAtividade nivelAtividade;
 
@@ -45,11 +42,10 @@ public class UsuarioPerfilDTO {
         this.plano = usuario.getPlano();
         this.dataUltimoTreino = usuario.getDataUltimoTreino();
 
-        // Popula os campos GOLD apenas se o plano for correspondente
         if (usuario.getPlano() == EPlano.GOLD) {
             this.idade = usuario.getIdade();
             this.peso = usuario.getPeso();
-            this.altura = usuario.getAltura();
+            this.altura = usuario.getAltura(); // Agora esta linha funciona corretamente
             this.objetivo = usuario.getObjetivo();
             this.nivelAtividade = usuario.getNivelAtividade();
         }
